Added CreatePost and EditPost page

+++ src/redux/postsSlice.js
@@ -14,10 +14,16 @@ const postsSlice = createSlice({
             localStorage.setItem('posts', JSON.stringify(state.posts));
         },
 
-        updatePost: (state, action) =>{
-            const index = state.posts.findIndex(p => p.id === action.payload.id);
-            state.posts[index] = action.payload;
-            localStorage.setItem('posts', JSON.stringify(state.posts));
+        updatePost: (state, action) => {
+            const { id, title, excerpt, content, thumbnail } = action.payload;
+            const existingPost = state.posts.find(post => post.id === id);
+            if (existingPost) {
+              existingPost.title = title;
+              existingPost.excerpt = excerpt;
+              existingPost.content = content;
+              existingPost.thumbnail = thumbnail;
+              localStorage.setItem('posts', JSON.stringify(state.posts));
+            }
         },
 
         deletePost: (state, action) =>{
